refactor(RunImage): basic tab
