+++ packages/renderer/src/lib/image/RunImage.spec.ts
@@ -520,8 +520,10 @@ describe('RunImage', () => {
     await new Promise(resolve => setTimeout(resolve, 600));
 
     const button = screen.getByRole('button', { name: 'Start Container' });
-    await tick();
-    expect((button as HTMLButtonElement).disabled).toBeTruthy();
+
+    await vi.waitFor(() => {
+      expect(button).toBeDisabled();
+    });
   });
 
   test('Expect able to play with devices', async () => {
